Final Commit
Ready for Review. Renamed the files to their proper names. Added in the scripts to the html page.
Added the HTML content to the page via JavaScript. A for loop looped through all of the candidates that were defined in a different script. The HTML was then set to the inside of an aside. A function that creates the HTML row for each candidate was created. A function that gets the div content for the bar was created. Added in comments to all of the files.